remove files of advancedDUngeon
