--- advancedDungeon/src/portal/portals/PortalColor.java
@@ -1,10 +1,0 @@
-package portal.portals;
-
-/** Helper enum to make the PortalFactory more flexible. */
-public enum PortalColor {
-
-  /** Represents the green portal. */
-  GREEN,
-  /** Represents the blue portal. */
-  BLUE
-}
