--- advancedDungeon/src/portal/portals/components/PortalIgnoreComponent.java
@@ -1,6 +1,0 @@
-package portal.portals.components;
-
-import core.Component;
-
-/** Holder class for all entities that should be ignored by the Portal. */
-public class PortalIgnoreComponent implements Component {}
