--- advancedDungeon/src/portal/util/AdvancedLevel.java
@@ -1,54 +1,0 @@
-package portal.util;
-
-import core.level.DungeonLevel;
-import core.level.utils.DesignLabel;
-import core.level.utils.LevelElement;
-import core.utils.Point;
-import java.util.Map;
-import java.util.Optional;
-
-/**
- * This class is used to store the values from a parsed level file. It contains the layout (the
- * tiles), the design label, the player start position and the custom points. This class is used in
- * the LevelParser.
- */
-public abstract class AdvancedLevel extends DungeonLevel {
-
-  private final DesignLabel designLabel;
-
-  /**
-   * Call the parent constructor of a tile level with the given layout and design label. Set the
-   * start tile of the player to the given heroPos.
-   *
-   * @param layout 2D array containing the tile layout.
-   * @param designLabel The design label for the level.
-   * @param namedPoints The custom points of the level.
-   * @param name The name of the level.
-   */
-  public AdvancedLevel(
-      LevelElement[][] layout,
-      DesignLabel designLabel,
-      Map<String, Point> namedPoints,
-      String name) {
-    super(layout, designLabel, namedPoints, name);
-    this.designLabel = designLabel;
-  }
-
-  @Override
-  public void onTick(boolean isFirstTick) {
-    if (isFirstTick) {
-      onFirstTick();
-    } else {
-      onTick();
-    }
-  }
-
-  /**
-   * Get the design of this level.
-   *
-   * @return The Design of this level.
-   */
-  public Optional<DesignLabel> designLabel() {
-    return Optional.ofNullable(designLabel);
-  }
-}
